+++ lib/dbf/opening.ex
@@ -149,7 +149,12 @@ defmodule DBF.Opening do
   defp find_memo_probe(resource, header, schema) do
     memo_fields =
       Enum.filter(schema.fields, fn field ->
-        field.decoder in [{:text, :memo}, {:binary, :text_memo}, {:binary, :binary_memo}]
+        field.decoder in [
+          {:text, :memo},
+          {:binary, :text_memo},
+          {:binary, :binary_memo},
+          {:binary, :picture_memo}
+        ]
       end)
 
     find_memo_probe(resource, header, memo_fields, 0)
@@ -199,6 +204,15 @@ defmodule DBF.Opening do
          _offset
        ) do
     {:ok, if(block == 0, do: nil, else: {block, :binary})}
+  end
+
+  defp parse_memo_probe(
+         <<block::little-unsigned-integer-size(32)>>,
+         _record_number,
+         %{decoder: {:binary, :picture_memo}},
+         _offset
+       ) do
+    {:ok, if(block == 0, do: nil, else: {block, :picture})}
   end
 
   defp parse_memo_probe(raw_pointer, record_number, field, offset) do